Fixed centralized wrapper

+++ src/rlenv/wrappers/penalty_wrapper.py
@@ -14,7 +14,7 @@ class TimePenalty(RLEnvWrapper):
         super().__init__(env)
         self.penalty = penalty
 
-    def step(self, action: npt.NDArray[np.int32]):
+    def step(self, action: npt.NDArray[np.int64]):
         obs, reward, *rest = self.wrapped.step(action)
         reward = [r - self.penalty for r in reward]
         return obs, reward, *rest
